audio: fix LP64 DSP-channel back-link truncation (unk8 u32->uintptr_t) + null-deref-before-check
TDSPChannel::getLogicalChannel returns (TChannel*)unk8, but unk8 was u32 and the channel ptr
was passed as (u32)(uintptr_t)this -> 64-bit host pointer TRUNCATED -> garbage ptr -> SIGSEGV
in updatecallDSPChannel once the DSP pipeline was actually driven. Widen the channel-ptr path
u32->uintptr_t (unk8, alloc/allocate/free sigs, drop (u32) casts). Also move
updatecallDSPChannel's 'mgr = channel->unk4' after the null check (null-deref-before-check).
uintptr_t==u32 on GC (faithful). Adds SB_DBG_AUDIO startSeq trace.

+++ include/JSystem/JAudio/JASystem/JASDSPChannel.hpp
@@ -24,8 +24,13 @@ public:
 
 	static void initAll();
 	static void updateAll();
-	static TDSPChannel* alloc(u32 param1, u32 param2);
-	static int free(TDSPChannel* channel, u32 param);
+	// param2/param carries a TChannel* (the logical-channel back-link stored in
+	// unk8). LP64: it MUST be pointer-width — the decomp typed it u32 (fine on
+	// 32-bit GC), which truncates a 64-bit host pointer and makes
+	// getLogicalChannel() return garbage -> SIGSEGV in updatecallDSPChannel once
+	// the DSP pipeline is actually driven (2026-07-17). uintptr_t == u32 on GC.
+	static TDSPChannel* alloc(u32 param1, uintptr_t param2);
+	static int free(TDSPChannel* channel, uintptr_t param);
 	TDSPChannel* getHandle(u32 handle);
 	u32 getNumUse();
 	u32 getNumFree();
@@ -33,7 +38,7 @@ public:
 	f32* getHistory();
 
 	void init(u8 param);
-	BOOL allocate(u32 param);
+	BOOL allocate(uintptr_t param); // LP64: holds a TChannel* (see alloc/free above)
 	void free();
 	void play();
 	void stop();
@@ -69,7 +74,7 @@ public:
 	/* 0x0 */ u8 unk3;
 	/* 0x4 */ u16 unk4;
 	/* 0x6 */ u16 unk6;
-	/* 0x8 */ u32 unk8;
+	/* 0x8 */ uintptr_t unk8; // LP64: holds a TChannel* back-link (was u32 = truncating on host)
 	/* 0xC */ DSPInterface::DSPBuffer* unkC;
 	/* 0x10 */ int (*unk10)(TDSPChannel*, u32);
 };
